refactor: turned navbar into header and navbar components and moved files around
