+++ src/app/(frontend)/_components/Footer.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Footer() {
+  return (
+    <div className="flex w-full justify-center bg-black text-white">
+      <div>© AUCC 2025</div>
+    </div>
+  )
+}
