--- src/app/(frontend)/components/Footer.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Footer() {
-  return <div>goodbye</div>
-}
